Updated lib and moved functions to seperate files
