+++ index.js
@@ -1,11 +1,3 @@
-const slash = function() {
-    if (process.platform.includes('win')) {
-        return '\\';
-    } else {
-        return '/';
-    }
-};
-
 /**
  * ### Logs the given message to the console following the given format:
  * `{TIME} | {MODULE} | {MESSAGE}`
@@ -17,70 +9,12 @@ const slash = function() {
  * @param {number} level The level of log message (0 = info, 1 = error, 2 = good news)
  */
 function log(message, level = 0) {
-    function getName(string) {
-        let name = string;
-        const dirs = [''];
-        dirs.shift();
-        for (const word of name.split(slash())) {
-            dirs.push(word.charAt(0).toUpperCase() + word.slice(1));
-        }
-        name = dirs
-            .join('/')
-            .replace('.js', '')
-            .replace('.ts', '')
-            .replace('.jsx', '')
-            .replace('.tsx', '');
-
-        return name;
-    }
-
-    const type = getName(
-        require.main.filename.replace(require('app-root-path').path + slash(), ''),
+    const type = require('./lib/getName').getName(
+        require.main.filename.replace(
+            require('app-root-path').path + require('./lib/slash').slash(),
+            '',
+        ),
     );
-
-    // #region Timestamp
-    function time() {
-        const date = new Date();
-
-        // #region Hour
-        let hour = date.getHours();
-
-        if (hour > 12) {
-            hour = hour - 12;
-        }
-
-        hour = hour.toString();
-
-        if (hour.length == 1) {
-            hour = '0' + hour.toString();
-        }
-        // #endregion Hour
-
-        // #region Minute
-        let minute = date.getMinutes();
-
-        minute = minute.toString();
-
-        if (minute.length == 1) {
-            minute = '0' + minute.toString();
-        }
-        // #endregion Minute
-
-        // #region second
-        let second = date.getSeconds();
-
-        second = second.toString();
-
-        if (second.length == 1) {
-            second = '0' + second.toString();
-        }
-        // #endregion second
-
-        const timeVar = `${hour.toString()}:${minute.toString()}:${second.toString()}`;
-
-        return timeVar;
-    }
-    // #endregion Timestamp
 
     const chalk = require('chalk');
 
@@ -101,9 +35,9 @@ function log(message, level = 0) {
     }
 
     console.log(
-        `${chalk.green(time())} | ${lvl} | ${chalk.cyanBright(
-            '[' + type + ']',
-        )} ${message}`,
+        `${chalk.green(
+            require('./lib/time').time(),
+        )} | ${lvl} | ${chalk.cyanBright('[' + type + ']')} ${message}`,
     );
 }
