+++ lib/slash.js
@@ -1,0 +1,7 @@
+module.exports.slash = function slash() {
+    if (process.platform.includes('win')) {
+        return '\\';
+    } else {
+        return '/';
+    }
+};
