+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.1.4",
   "description": "A simple, clean, javascript logger",
   "main": "index.js",
-  "types": "./typings/index.d.ts",
+  "types": "./typings/",
   "scripts": {
     "test": "npm run jstest && npm run tstest",
     "jstest": "node ./tests/JSTest.js",
