+++ typings/lib/getName.d.ts
@@ -1,0 +1,1 @@
+export function getName(string: any): any;
