+++ typings/lib/slash.d.ts
@@ -1,0 +1,1 @@
+export function slash(): "\\" | "/";
