+++ typings/lib/time.d.ts
@@ -1,0 +1,1 @@
+export function time(): string;
